[Embedded AR] - added the option to select resolution of 480p24 because at least the VU cannot display 480p24. This format is quite common on US DVDs or rips. - only files with width between 700 and 720 and height less or equal 480 and frequency of 24,30,60Hz are played as 480 - only files with width between 700 and 720 and height less or equal 576 and frequency of 25,50Hz are played as 576 - only files with width of 1280 and height less or equal 720 are played as 720 - all other will be played in the configured resolution
These changes avoid errors which caused 16:9 576 videos played as 480 and 16:9 1280 videos played as 576. Because  the width is used to detect the mode double scaling will be avoid most (all) of the times. 320x240 for instance will be scaled to the configured video mode and not a mode from auto resolution